scaffold new stages and types
